app restructured again
