+++ hw12_13_14_15_16_calendar/internal/app/calendar/defaults.go
@@ -1,0 +1,9 @@
+package app
+
+import "time"
+
+// expectedFields is a map of expected configuration fields and their default values.
+var expectedFields = map[string]any{
+	"retries":       int(0),
+	"retry_timeout": time.Duration(0),
+}
